requirements.txt
Flask
gunicorn
python-binance
pytz
flask_cors
requests